+++ CHANGES.md
@@ -13,6 +13,7 @@
 * Runtime: Implement buffer for in_channels
 * Lib: wheel event binding
 * Test: track external used in the stdlib and unix
+* Runtime: add support for unix_opendir, unix_readdir, unix_closedir, win_findfirst, win_findnext, win_findclose
 
 ## Bug fixes
 * Compiler: fix rewriter bug in share_constant (fix #1247)